Moved and renamed SelectYear and SelectTreatmentUnits code
