+++ src/components/RegisterPage/main_component.tsx
@@ -7,7 +7,7 @@ import LEGEND from "../TargetLevels";
 import { MedicalFields } from "../MedicalFields";
 import { IndicatorTable } from "../IndicatorTable";
 import { useMedicalFieldsQuery } from "../../helpers/hooks";
-import { OptsTu } from "../select_multi";
+import { OptsTu } from "../SelectTreatmentUnits";
 
 interface AggData {
   nation: {
